21-1 final-term problem solving
The requirement was sorting of a given list consisting of words which are combined with digit or not. This source code is how I solve the problem that I couldn't deal with during final-term.
Through this problem, I have learned the importance of knowing how to implement 'sorting'. I have used 'sort()' familiarly and neglected understanding the algorithm of the function. There are 3 basic style of sorting - selection sort, insertion sort, and bubble sort. The third one is what I utilized editing my source code.